feat: migrate agent settings to Agent page, remove Agent tab from Settings (#279)
