+++ inc/Core/Admin/shared/queries/tools.js
